refactor: single script that manages all pages
