+++ package.json
@@ -15,7 +15,7 @@
     "prodChrome": "npm run licenseList && npm run chrome && node manage/prodChanges.js && npm run zip",
     "prodFirefox": "npm run licenseList && npm run firefox && node manage/prodChanges.js && npm run zip",
     "--- Build scripts": "---",
-    "build": "node manage/initBuild.js && esbuild src/scripts/background/background.js --bundle --outdir=dist && esbuild src/scripts/content/tab.js --bundle --outdir=dist && esbuild src/scripts/pages/options.js --bundle --outdir=dist/pages",
+    "build": "node manage/initBuild.js && esbuild src/scripts/background/background.js --bundle --outdir=dist && esbuild src/scripts/content/tab.js --bundle --outdir=dist && esbuild src/scripts/pages/pages.js --bundle --outdir=dist/pages",
     "build:l": "npm run licenseList && npm run build",
     "build:w": "nodemon --exec npm run build --watch",
     "watchAll": "nodemon --exec npm run all --watch",
