+++ src/scripts/pages/options.html
@@ -3,11 +3,11 @@
 <head>
   <meta charset="UTF-8">
   <title>Save my Chatbot Options</title>
-  <link rel="stylesheet" href="options.css">
+  <link rel="stylesheet" href="pages.css">
 </head>
 <body>
 <h1>Options de l'Extension</h1>
-  <div id="options-container"></div>
-  <script src="options.js"></script>
+<div id="options-container" class="container"></div>
+<script src="pages.js"></script>
 </body>
 </html>
