+++ src/scripts/pages/shoelace-styles.js
@@ -1,0 +1,2 @@
+import '@shoelace-style/shoelace/dist/themes/light.css';
+import '@shoelace-style/shoelace/dist/themes/dark.css';
